1. aggregate string metrics by concatenating as a CSV string 2. don't call doubleValue since we want the Double object

+++ hawkular-wildfly-agent/src/main/java/org/hawkular/agent/monitor/storage/NumericMetricDataPoint.java
@@ -45,7 +45,7 @@ public class NumericMetricDataPoint extends MetricDataPoint {
         if (v instanceof Number) {
             return ((Number) v).doubleValue();
         } else {
-            return (v != null) ? (Double.valueOf(v.toString())).doubleValue() : Double.valueOf(0.0);
+            return (v != null) ? (Double.valueOf(v.toString())) : Double.valueOf(0.0);
         }
     }
 }
